Don't sync files already synced. Fix #1

+++ src/main/scala/rewind/stores/GribStore.scala
@@ -2,6 +2,7 @@ package rewind.stores
 
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
+import scala.util.{Left, Right}
 
 import cats.implicits._
 import cats.effect.IO
@@ -24,26 +25,41 @@ class GribStore(httpClient: Client[IO], storageConf: ObjectStorage) {
       .withRegion(storageConf.region)
       .build()
 
-  def syncAt(date: LocalDate, hour: Int): IO[String] = {
+  def syncAt(date: LocalDate, hour: Int, force: Boolean): IO[String] = {
     val noaaUri = GribStore.noaaUri(date, hour)
-    for {
-      _ <- IO(logger.info(s"Downloading NOAA file on $date at $hour: $noaaUri"))
-      content <- httpClient.expect[String](noaaUri)
-      filename = GribStore.filename(date, hour)
-      _ <- IO(logger.info(s"Download successful, uploading to $filename"))
-      _ <- IO(s3.putObject(storageConf.bucket, filename, content))
-      _ <- IO(logger.info(s"Upload successful to $filename"))
-    } yield filename
+    exists(date, hour).flatMap { alreadyThere =>
+      val filename = GribStore.filename(date, hour)
+      if (alreadyThere && !force) {
+        IO(logger.info(s"Skipping, file already synced")).map(_ => filename)
+      } else {
+        for {
+          _ <- IO(
+            logger.info(s"Downloading NOAA file on $date at $hour: $noaaUri"))
+          content <- httpClient.expect[String](noaaUri)
+          _ <- IO(logger.info(s"Download successful, uploading to $filename"))
+          _ <- IO(s3.putObject(storageConf.bucket, filename, content))
+          _ <- IO(logger.info(s"Upload successful to $filename"))
+        } yield filename
+      }
+    }
   }
 
-  def syncOn(date: LocalDate): IO[List[String]] = {
+  def syncOn(date: LocalDate, force: Boolean): IO[List[String]] = {
     for {
       _ <- IO(logger.info(s"Starting sync on $date"))
       names <- 0.to(3).map(_ * 6).toList.traverse { hour =>
-        syncAt(date, hour)
+        syncAt(date, hour, force)
       }
       _ <- IO(logger.info(s"Finished sync on $date"))
     } yield names
+  }
+
+  def exists(date: LocalDate, hour: Int): IO[Boolean] = {
+    IO(s3.getObjectMetadata(storageConf.bucket, GribStore.filename(date, hour))).attempt
+      .map {
+        case Left(_)  => false
+        case Right(_) => true
+      }
   }
 }
